0.95 Version 08.17.2014

+++ src/com/yuanwei/resistance/gridviewtest/SquareLayout.java
@@ -32,7 +32,7 @@ public class SquareLayout extends RelativeLayout {
         int childHeightSize = getMeasuredHeight();
         //
          widthMeasureSpec = MeasureSpec.makeMeasureSpec(childWidthSize, MeasureSpec.EXACTLY);
-         heightMeasureSpec = (int)(widthMeasureSpec);
+         heightMeasureSpec = (widthMeasureSpec);
         super.onMeasure(widthMeasureSpec, heightMeasureSpec);
     }
 }
